util/render: move progress writer from cli/cmds/provisioning
