+++ internal/cli/cmds/provisioning/cluster_artifact.go
@@ -253,7 +253,7 @@ func (c *cmdClusterArtifactGetArchive) run(cmd *cobra.Command, args []string) er
 	quiet, _ := cmd.Flags().GetBool("quiet")
 	format := fmt.Sprintf("Fetching artifact %q archive for cluster %q: %%s", artifactName, clusterName)
 
-	progress, writer := progressWriter(targetFile, format, quiet)
+	progress, writer := render.ProgressWriter(targetFile, format, quiet)
 
 	size, err := file.SafeCopy(writer, archiveReader)
 	if err != nil {
@@ -321,7 +321,7 @@ func (c *cmdClusterArtifactGetFile) run(cmd *cobra.Command, args []string) error
 	quiet, _ := cmd.Flags().GetBool("quiet")
 	format := fmt.Sprintf("Fetching file %q for artifact %q of cluster %q: %%s", filename, artifactName, clusterName)
 
-	progress, writer := progressWriter(targetFile, format, quiet)
+	progress, writer := render.ProgressWriter(targetFile, format, quiet)
 
 	size, err := file.SafeCopy(writer, fileReader)
 	if err != nil {
