+++ internal/cli/cmds/provisioning/token.go
@@ -442,7 +442,7 @@ func (c *cmdTokenGetImage) run(cmd *cobra.Command, args []string) (err error) {
 	quiet, _ := cmd.Flags().GetBool("quiet")
 	format := fmt.Sprintf("Fetching image for token %s: %%s", id)
 
-	progress, writer := progressWriter(targetFile, format, quiet)
+	progress, writer := render.ProgressWriter(targetFile, format, quiet)
 
 	size, err := file.SafeCopy(writer, imageReader)
 	if err != nil {
